Fixed a where buildings adds double destruction points
Some ammonitions are hitting with multiple damage types (like direct hit damage and indirect hit damage). If the building is already destroyed by the direct hit damage which occures first the building will add its points again if the indirect damage hits.

To avoid that I added a check, if the current damage is already above 1 and  if so let the EH exit as it is no longer needed