Started groundwork for roadblock and city creation
This commit has gone a bit crazy, sorry Spoffy

Started work on a script that handles the creation of roadblock objects for AI faction. Had to stop to change the parameters for cycleSpawn to accept more vehicles and groups.

Changed createAISite to handle AI and Rebel roadblock creation internal and adjusted the parameters for cycleSpawn. Adjusted the priority for the debug message to 2.

Added file for rebel roadblock creation, currently empty.

Changed parameters for cycleSpawn and adjusted the priority of the logs to 2.

Added createRebelControlPoint to the functions.hpp