created selector_new.sqf
The integration of the MIE has massively changed the template system.
Parts of that were also the mod-detection variables as well as how mods are loaded, including the logistic nodes.

I have adapted what I could, commented out stuff that is not Vanilla, RHS or 3CB and hope that I didn't fuck up too much.